move schemas to types and infer from zod
